+++ RELEASES.md
@@ -8,6 +8,7 @@ as-is, without any warranty. -->
 Version 1.22.0 (unreleased)
 ===========================
 
+  * String parsing now supports an optional exponent.
   * The [`lit`][f-l-1-22] method was added to all fixed-point numbers.
 
 [f-l-1-22]: https://docs.rs/fixed/~1.22/fixed/struct.FixedI32.html#method.lit